First Version
This version has been tested with a 20x20 maze. The walls have been initialized randomly. The problem of this implementation is that all the code is badly written and disorganized inside the class PathFinder.java. Future versions will aim to separate those functions into more, simpler clases, in order to make it more easily understandable.